requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
pyserial==3.5
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
pyjwt==2.8.0
email-validator==2.1.0 